almost entirely separated grid from game
