+++ game/game.go
@@ -2,7 +2,6 @@ package game
 
 import (
 	"errors"
-	"strings"
 )
 
 var (
@@ -10,48 +9,19 @@ var (
 	ErrBadGridChar  = errors.New("Characters in grid must either be - or *")
 )
 
+
 type Game struct {
-	width, height int
 	grid          grid
 }
 
-func NewGame(width, height int) *Game {
+func NewGame(grid grid) *Game {
 	g := new(Game)
-	g.width = width
-	g.height = height
-
-	g.grid = newGrid(g.width, g.height)
-
+	g.grid = grid
 	return g
 }
 
-func NewGameFromString(in string) (*Game, error) {
-	rows := strings.Split(strings.TrimSpace(in), "\n")
-	height := len(rows)
-	width := len(rows[0])
-
-	g := NewGame(width, height)
-
-	for rowNumber, row := range rows {
-		if len(row) != width {
-			return nil, ErrBadGridShape
-		}
-		for colNumber, col := range row {
-			if col == '*' {
-				g.grid[rowNumber][colNumber] = true
-			} else if col == '-' {
-				g.grid[rowNumber][colNumber] = false
-			} else {
-				return nil, ErrBadGridChar
-			}
-		}
-	}
-
-	return g, nil
-}
-
 func (g *Game) Step() {
-	nextGrid := newGrid(g.height, g.width)
+	nextGrid := g.grid.cleanClone()
 
 	for _, cell := range g.grid.getCells() {
 		x := cell.x
