+++ game/game_test.go
@@ -2,96 +2,14 @@ package game
 
 import "testing"
 
-func TestInitialisesAnEmptyGrid(t *testing.T) {
-	width := 5
-	height := 8
-	game := NewGame(width, height)
-
-	if game.grid[height-1][width-1] != false {
-		t.Error("Expected bottom right corner to be set to false")
-	}
-
-	if len(game.grid) != height {
-		t.Error("Incorrect value for height, got", len(game.grid), "expected", height)
-	}
-
-	if len(game.grid[0]) != width {
-		t.Error("Incorrect value for width, got", len(game.grid[0]), "expected", width)
-	}
-}
-
-func TestItCanCreateGridFromString(t *testing.T) {
-	input := `
--*-
----
----
-`
-	game, _ := NewGameFromString(input)
-
-	if game == nil {
-		t.Error("Expected a grid to be returned but got nil")
-	}
-
-	if game.grid[0][1] != true {
-		t.Error("Expected grid item to be set to true but it wasnt", game.grid)
-	}
-
-}
-
-func TestItCanCreateGridFromStringAgain(t *testing.T) {
-	input := `
-*--
--*-
---*
-`
-	game, _ := NewGameFromString(input)
-
-	if game == nil {
-		t.Error("Expected a grid to be returned but got nil")
-	}
-	if !(game.grid[0][0] && game.grid[1][1] && game.grid[2][2]) {
-		t.Error("Expected grid item to be set to true but it wasnt", game.grid)
-	}
-
-	if game.grid[0][1] && game.grid[0][2] &&
-		game.grid[1][0] && game.grid[1][2] &&
-		game.grid[2][0] && game.grid[2][1] {
-		t.Error("Expected grid item to be set to false but it wasnt", game.grid)
-	}
-
-}
-
-func TestItCanOutputStringGrid(t *testing.T) {
-	input := `
----
----
----
-`
-	game, _ := NewGameFromString(input)
-
-	game.grid[0][0] = true
-
-	expected := `
-*--
----
----
-`
-	if game.grid.String() != expected {
-		t.Log("Expected")
-		t.Log(expected)
-		t.Log("Got")
-		t.Log(game.grid.String())
-		t.Error("Grid was not outputted as expected")
-	}
-}
-
 func TestFewerThanTwoLiveCellsDiesOnGeneration(t *testing.T) {
 	input := `
 ---
 -*-
 ---
 `
-	game, _ := NewGameFromString(input)
+	grid, _ := NewGridFromString(input)
+	game := NewGame(grid)
 
 	game.Step()
 
@@ -116,7 +34,8 @@ func TestTwoByTwoSquareLivesToNextGeneration(t *testing.T) {
 -**
 -**
 `
-	game, _ := NewGameFromString(input)
+	grid, _ := NewGridFromString(input)
+	game := NewGame(grid)
 
 	game.Step()
 
@@ -141,7 +60,8 @@ func TestTwoByDiagonalTripleDiesExceptCenterCell(t *testing.T) {
 -*-
 *--
 `
-	game, _ := NewGameFromString(input)
+	grid, _ := NewGridFromString(input)
+	game := NewGame(grid)
 
 	game.Step()
 
@@ -166,7 +86,8 @@ func TestCenterCellWith4NeighboursDies(t *testing.T) {
 ***
 *-*
 `
-	game, _ := NewGameFromString(input)
+	grid, _ := NewGridFromString(input)
+	game := NewGame(grid)
 
 	game.Step()
 
@@ -191,7 +112,8 @@ func TestDeadCenterCellWith3NeighboursResurrects(t *testing.T) {
 ---
 *--
 `
-	game, _ := NewGameFromString(input)
+	grid, _ := NewGridFromString(input)
+	game := NewGame(grid)
 
 	game.Step()
 
@@ -207,39 +129,5 @@ func TestDeadCenterCellWith3NeighboursResurrects(t *testing.T) {
 		t.Log("Got")
 		t.Log(result)
 		t.Error("Grid was not outputted as expected")
-	}
-}
-
-func TestItRejectsNonRectangularGrids(t *testing.T) {
-	input := `
--*-----
----
----
-`
-	_, err := NewGameFromString(input)
-
-	if err == nil {
-		t.Error("Expected an error")
-	}
-
-	if err != ErrBadGridShape {
-		t.Error("Expected", ErrBadGridShape, "but got", err)
-	}
-}
-
-func TestItRejectsUnexpectedCharsInGrids(t *testing.T) {
-	input := `
----
--2-
----
-`
-	_, err := NewGameFromString(input)
-
-	if err == nil {
-		t.Error("Expected an error")
-	}
-
-	if err != ErrBadGridChar {
-		t.Error("Expected", ErrBadGridChar, "but got", err)
 	}
 }
